hal: microchip: mec5: Add VWire write timeout mechanism
Target to Controller virtual wire changes are not automatically
transmitted to the eSPI Host controller. The target sets VWires
available hardware status. If the eSPI alert mechanism is supported
by the target and configured by the host controller the target will
assert eSPI alert in any eSPI HW status changes. The host sees
the alert, issues a GET_STATUS command, processes the status bits
and issues a GET_VWIRES command.  When the target finishes
processing GET_VWIRES it will clear its read-only Target-to-Controller
VWire change status bits. Based on eSPI clock frequency and bus
width this process could take up to 100 us. We add some API's for the
eSPI driver to pass a structure containing a function pointer to a
delay routine, parameter to pass to the delay routine and a loop count.
If the data structure and its function poiner are valid the HAL will
poll the Target-to-Controller VWire changes bits for the VWire group
until all bits are clear or timeout.

Signed-off-by: Scott Worley <scott.worley@microchip.com>