hal: microchip: mec5: I2C bug fixes idle interrupt and start
Microchip MEC5 HAL I2C idle interrupt enable routine was using
the defines for the idle status bit in the completion register
instead of the defines for the enable bit in the configuration
register. Added code to clear the idle status before writing
the controller mode command register which begins the
transaction.

Signed-off-by: Scott Worley <scott.worley@microchip.com>