hal: microchip: mec5: Fix EXTLEN bug in I2C v3.7 register header
The I2C-SMB extended length register is not correctly documented
in the data sheet. The correct register bit postions and masks
were added to the v3.8 I2C-SMB header but not to the v3.7.
Both version implement the same EXTLEN register. The v3.7 header
EXTLEN register defines were updated to match the hardware.

Signed-off-by: Scott Worley <scott.worley@microchip.com>