hal: microchip: sam: Update SAMA7G5 FLEXCOM USART/SPI/TWI definitions
In Zephyr SAME70 and SAMV71 was supported already. The driver for USART
could be reused for SAMA7G5 series. As there's no stand-alone USART
peripheral in SAMA7G5, the USART function offered by FLEXCOM can be
used instead.
Here add some definitions for USART which are ported from "flexcom.h".
It is similar for SPI and TWI.

Signed-off-by: Tony Han <tony.han@microchip.com>